refactor: split workflows/tracker.py into package
Move dag/gate/sub-workflow/fan-out helpers out of the 875-line
WorkflowTracker class into focused modules. Tracker stays the
orchestrator; helpers are free functions taking the tracker as the
first argument. No behavior changes.
